back_end_sdp/xdd/requirements.txt
Django==5.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.4
django-cors-headers==4.2.0